+++ packages/components/src/components/DiagramSequence.vue
@@ -1,11 +1,11 @@
 <template>
-  <div class="sequence-diagram" id="sequence-diagram-ui">
+  <div class="sequence-diagram" id="sequence-diagram-ui" :key="renderKey">
     <template v-for="(actor, index) in actors">
       <VActor
         :actor="actor"
         :row="1"
         :index="index"
-        :height="diagramSpec.rowCount"
+        :height="diagramSpec.actions.length"
         :selected-actor="selectedActor"
       />
     </template>
@@ -15,6 +15,7 @@
           :actionSpec="action"
           :selected-actions="selectedActions"
           :collapsed-actions="collapsedActions"
+          :selected-trace-event="selectedTraceEvent"
         />
       </template>
       <template v-if="action.nodeType === 'return'">
@@ -62,12 +63,24 @@ export default {
     serializedDiagram: {
       type: Object,
     },
+    selectedTraceEvent: {
+      type: Object,
+      default: null,
+    },
     selectedEvents: {
       type: Array,
     },
   },
+
+  data() {
+    return {
+      renderKey: 0,
+    };
+  },
+
   computed: {
     diagram() {
+      this.renderKey += 1;
       let result: Diagram | undefined;
       if (this.serializedDiagram) {
         result = unparseDiagram(this.serializedDiagram as Diagram);
@@ -113,6 +126,21 @@ export default {
         ...(codeObject as any).ancestors().map((ancestor: CodeObject) => ancestor.fqid),
       ];
       return this.actors.find((actor) => ancestorIds.indexOf(actor.id) !== -1);
+    },
+  },
+
+  methods: {
+    focusFocused() {
+      setTimeout(() => {
+        const element = this.$el.querySelector(
+          '.call.focused .self-call, .call.focused .call-line-segment'
+        );
+        if (!element) {
+          return;
+        }
+
+        element.scrollIntoView({ behavior: 'smooth', block: 'center' });
+      }, 16);
     },
   },
 };