+++ packages/components/src/components/sequence/DiffChannel.vue
@@ -16,7 +16,7 @@
 </template>
 
 <script lang="ts">
-//@ts-nocheck
+// @ts-nocheck
 import { DiffMode } from '@appland/sequence-diagram';
 import { ActionSpec } from './ActionSpec';
 
@@ -42,8 +42,8 @@ export default {
     },
     diffRowLabel(): string {
       if (this.actionSpec.action.diffMode === DiffMode.Insert) return '+';
-      else if (this.actionSpec.action.diffMode === DiffMode.Delete) return '-';
-      else if (this.actionSpec.action.diffMode === DiffMode.Change) return '+/-';
+      if (this.actionSpec.action.diffMode === DiffMode.Delete) return '-';
+      if (this.actionSpec.action.diffMode === DiffMode.Change) return '+/-';
 
       return '';
     },