+++ packages/components/src/store/vsCode.js
@@ -58,10 +58,12 @@ export function buildStore() {
         });
       },
 
+      // Show information about a code object in the sidebar.
+      // The code object can be a package, class, function, SQL, etc, or it can
+      // be a specific event.
       [SELECT_CODE_OBJECT](state, selection) {
-        if (!Array.isArray(selection)) selection = [selection];
-
-        state.selectionStack.push(...selection);
+        let selectionStack = Array.isArray(selection) ? selection : [selection];
+        state.selectionStack.push(...selectionStack);
         state.selectedLabel = null;
         state.selectedTraceEvent = null;
       },
@@ -86,6 +88,7 @@ export function buildStore() {
         state.currentView = view;
       },
 
+      // Show the selected event in the current diagram (not the sidebar).
       [SET_SELECTED_TRACE_EVENT](state, event) {
         state.selectedTraceEvent = event;
       },