+++ packages/models/types/index.d.ts
@@ -97,8 +97,9 @@ declare module '@appland/models' {
     readonly id: string;
     readonly name: string;
     readonly type: CodeObjectType;
+    readonly data: any;
     readonly static: boolean;
-    readonly location: string;
+    readonly location: string | undefined;
     readonly locations: string[];
     readonly packageOf: string;
     readonly classOf: string;
@@ -131,6 +132,7 @@ declare module '@appland/models' {
     readonly codeObjects: CodeObject[];
     readonly httpObject?: CodeObject;
     readonly sqlObject?: CodeObject;
+    readonly roots: CodeObject[];
 
     visit(fn: (codeObject: CodeObject) => void): void;
     search(query: string): CodeObject[];